ref: divide notification code
