+++ packages/server/tsconfig.json
@@ -33,7 +33,7 @@
       "@validators/*": ["validators/*"],
       "*": ["types/*"]
     },
-    "outDir": "dist",
+    "outDir": "dist"
   },
   "ts-node": {
     "require": ["tsconfig-paths/register"]
